third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.request-header.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Request Header seen on all popin navigations/redirects
  assert_equals: expected "Initial(partitioned)-HTTP(partitioned)-JS(partitioned)-fetch(missing)-iframe(missing)-" but got "Initial(missing)-HTTP(missing)-JS(missing)-fetch(missing)-iframe(missing)-"
Harness: the test ran to completion.
